Create types.ts for swap reducer
